Add acceptance tests and a non-stacking middleware
Includes a non-stacking version of the middleware. The crucial different
is that the counter increment is performed before calling the next
request handler, which means that there is less risk of multiple
parallel requests that are hitting the same limit being handled when the
limit is actually exhausted.

We also modify the logic to skip rate limiting when the limit key is
`nil`. This allows limit implementations to return `nil` for requests
that do not 'match' the limit implementation. An example would be a HTTP
method limit that applies only to one method (such a limit was
implemented for the tests).