Pass in req to all RateLimit fns
We used to provide the request only to the `get-key` fn for generating
the rate limit key, but it turns out that we should really pass the
request to at least the `get-quota` fn as well so that we can
dynamically generate different quotas based on the request (eg we can
return a user-specific quota by looking up the user from the request and
the looking up the user-specific quota from a database).

Since we're at it, we might as well pass the request to the `get-period`
fn. This might not be quite as useful as the other two fns, but it will
allow the RateLimit author to dynamically generate quota periods - eg
one user could have a rate limit of 100req/min and another 10req/s.