Add :test-selectors
Leiningen :test-selectors allows a dev to just run a subset of the
tests. In our case we want to allow a dev to skip the Redis tests in
case they don't have a redis sever available.

At the moment we will ignore all of the acceptance tests, because they
run against Redis. Ideally we could still run the aceptance tests
against the local storage backend and any future backends we might
introduce.